fix `iai_callgrind` metrics
- Fixed an existing bug where existing `CALLGRIND` tool metrics were using an incorrect name/slug from `iai`.
- add metrics from `DHAT` tool
- Fixed the `iai_callgrind` parser to account for either one or mulitple tools running, and added some tests for the new format.
- made sure all three adapters (`rust`, `iai-callgrind`, `magic`) are testing the different conditions for completeness.

Note: I consolidated all the measures into a `mod` hierarchy, to avoid name collisions.
The new consolidated hierarchy can elimiate a lot of the boiletplate code by reusing the generic trait directly.
For example, the validators in `lib/bencher_adapter/src/adapters/mod.rs`, or the wrappers `IaiMeasure` and `IaiCallgrindMeasure`.
But I didn't want to change too much at once, so I left it as is for now. Just pointing this out for future reference.